+++ ricaun.RevitTest.TestAdapter/Models/RunSettingsModel.cs
@@ -24,7 +24,7 @@
             public bool Close { get; set; }
 
             [XmlElement("Timeout")]
-            public int Timeout { get; set; }
+            public double Timeout { get; set; }
 
             [XmlElement("Verbosity")]
             public int Verbosity { get; set; }
